Creati controlli admin sulla modifica e cancellazione

+++ Spring/src/main/java/it/pw/controller/AdminUtenteFormController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import it.pw.dao.OrdiniDao;
 import it.pw.dao.UtenteDao;
 import it.pw.service.UtenteService;
 
@@ -20,10 +21,13 @@ public class AdminUtenteFormController {
 	@Autowired
 	UtenteDao utenteDao;
 	
+	@Autowired
+	OrdiniDao ordinidao;
+	
 	@GetMapping
 	String getPage(Model model) {
 		model.addAttribute("utenti",utenteDao.leggiTutti());
-		
+		model.addAttribute("esitoRegistrazione",true);
 		
 		
 		return "admin-utenti";
@@ -34,6 +38,13 @@ public class AdminUtenteFormController {
 	String eliminaUtente(HttpServletRequest request, HttpSession session, Model model) {
 		int id = 0;		
 		id= Integer.parseInt(request.getParameter("id"));
+		
+		if(ordinidao.confrontaDataUtente(utenteDao.getutenteById(id))) {
+			model.addAttribute("esitoRegistrazione",false);
+			model.addAttribute("utenti",utenteDao.leggiTutti());
+			return "admin-utenti";
+		}
+		
 		utenteDao.delete(utenteDao.getutenteById(id));
